+++ src/main/java/com/expediagroup/dropwizard/resilience4j/configuration/retry/RetryConfiguration.java
@@ -1,8 +1,7 @@
 package com.expediagroup.dropwizard.resilience4j.configuration.retry;
 
+import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-
-import org.hibernate.validator.constraints.NotEmpty;
 
 import com.fasterxml.jackson.annotation.JsonProperty;